+++ src/machinable/config/parser.py
@@ -1,103 +1,11 @@
-import importlib
-import inspect
-
 import regex
 from flatten_dict import unflatten
 
-from ..registration import Registration
 from ..utils.dicts import get_or_fail, read_path_dict, update_dict
-from ..utils.formatting import exception_to_str, msg
+from ..utils.formatting import msg
+from ..utils.importing import ModuleClass
 from ..utils.utils import is_valid_variable_name
 from .mapping import _reserved_keys, _used_keys
-
-
-class ModuleClass(object):
-    def __init__(self, module_name, args=None, baseclass=None):
-        self.module_name = module_name
-        self.args = args
-        self.baseclass = baseclass
-        self.default_class = None
-
-    def load(self, instantiate=True, default=None):
-        if default is None:
-            default = self.default_class
-
-        # allow overrides
-        registration = Registration.get()
-        on_before_component_import = registration.on_before_component_import(
-            module=self.module_name, baseclass=self.baseclass, default=default
-        )
-        if on_before_component_import is not None:
-            return on_before_component_import
-
-        module_class = None
-        try:
-            module = importlib.import_module(self.module_name)
-            try:
-                # reload if we are in interactive environments like jupyter
-                get_ipython().__class__.__name__
-                importlib.reload(module)
-            except NameError:
-                pass
-
-            for candidate, class_ in inspect.getmembers(module, inspect.isclass):
-                if self.baseclass is not None and not issubclass(
-                    class_, self.baseclass
-                ):
-                    continue
-
-                if class_.__module__ == self.module_name:
-                    module_class = class_
-                    break
-
-            if module_class is None:
-                module_class = AttributeError(
-                    f"Could not load module class from module '{self.module_name}'. "
-                    f"Make sure the module contains a class that inherits from "
-                    f"the baseclass 'machinable.{self.baseclass.__name__}'"
-                )
-        except ImportError as e:
-            if default is None:
-                module_class = ImportError(
-                    f"Could not import module '{self.module_name}' "
-                    f"that is specified in the machinable.yaml. "
-                    f"The following exception occurred: {exception_to_str(e)}. "
-                    f"If the module is a directory, consider creating an __init__.py."
-                )
-
-        on_component_import = registration.on_component_import(
-            component_candidate=module_class,
-            module=self.module_name,
-            baseclass=self.baseclass,
-            default=default,
-        )
-        if on_component_import is not None:
-            return on_component_import
-
-        if isinstance(module_class, (ImportError, AttributeError)):
-            raise module_class
-
-        if module_class is None:
-            module_class = default
-
-        if not instantiate:
-            return module_class
-
-        if self.args is None:
-            return module_class()
-
-        return module_class(**self.args)
-
-    def name(self):
-        class_ = self.load(instantiate=False)
-        return class_.__name__
-
-    def __call__(self, *args, **kwargs):
-        class_ = self.load(instantiate=False)
-        return class_(*args, **kwargs)
-
-    def __repr__(self):
-        return f"<machinable.config.ModuleClass(module_name={self.module_name}, args={self.args} baseclass={self.baseclass})>"
 
 
 def parse_mixins(config, valid_only=False):