+++ src/machinable/console/cli.py
@@ -1,5 +1,4 @@
 import click
-
 import pkg_resources
 
 from .execute import execute, execution