+++ src/machinable/execution/schedule.py
@@ -1,9 +1,9 @@
 import copy
 import random
 
-from ..config.parser import ModuleClass
 from ..core.component import Component as BaseComponent
 from ..utils.identifiers import encode_experiment_id, generate_component_id
+from ..utils.importing import ModuleClass
 from ..utils.traits import Jsonable
 from ..utils.utils import generate_seed